Lots of fixing for x64 compatibility. Fixed various str-related hangs due to using unsigned type. Changed version to 1.91 instead of 1.09 (prevent confusion)

+++ src/event.h
@@ -42,16 +42,16 @@
 			std::string msg;
 			//char *msg;
 			union {
-				int idata1;
+				size_t idata1;
 				double fdata1;
 				Socket socket;
 			};
 			union {
-				int idata2;
+				size_t idata2;
 				double fdata2;
 			};
 			union {
-				int idata3;
+				size_t idata3;
 				double fdata3;
 			};
